CurrentRecordingId is backed by the db. And we can send Location from db to Activity.
CurrentRecordingManager manages an append-only table that determines the most recent id we want to reply. In our test Activity, we can subscribe to the MockerService and can receive updates from the service. We're currently using mocked location data because the I can't get GPS for whatever reason, so this is pretty meta but I know the recording functionality works. So this is really cool. Just have to implement it on the platform.