We can send Locations to LocationManager subscribers.
We have a Thread that handles the MockLocation updates and timing,
which then broadcasts the update to every subscribed client in a
new one-off Thread. Right now I'm only really testing for
locationChanged, but by recording more details (such as providerChanged
or what have you).

Currently, we always say we are NOT replaying. However, that functionality
can work by just sending the actual value of isReplaying.

There's a Replay button now that kicks off the replaying process.
It works for the package 'android' but, I believe, it causes the
Android Fused Location Provider to crash. At 4:20am, though, I don't
care that much, so that's a TODO for later (granted, a very important
one).