LocationManager can listen for mocked updates.
I'm not really sure how to handle versioning since every time I
try to version the entire $ROOT/frameworks/base repo the remote
push fails. Ah, well, this will due for now because I need to have
multiple copies at this point.

We subscribe to MockerService (and create it if it is not created)
on construction. We have our own Handler for receiving updates
from the MockerService (which contains a Bundle that, essentially,
represents an android.location.Location and has some other metadata).

We also wrapped the ListenerTransport _handleMessage with an
isReplaying clause, which will stop all REAL changes from going to
the client app (or so I think?).

TODO: In our incoming handler, we need to broadcast the update to
all of our LocationListeners. We can just use the mListeners keySet
or so I think and just pass the values like _handleMessage. Nbd.

Night.