Started working on the MockSensorEventManager.
Turns out, we don't really need to care when we push
out sensor events because of:

"The rate sensor events are delivered at. This is only a hint to the system.
Events may be received faster or slower than the specified rate. Usually events
are received faster."

From the docs for SensorManager.registerListener