We can push out mock sensor data.
TODO: Fix But the null-pointer bug in the MockerService when
a location or sensor has only one entry.

In MainActivity, added a TextView to show the status of the
sensor recording. We only record sensor data when we have location
data because idk. And now the Overwrite Dialog doesn't show up
while we are recording.

In MockerService, we separated clients into location and sensor
clients and created a SensorReplayer thread that is based on
LocationReplayer.

MockSensorEvent has a new constructor.

Helper functions in MockSensorEventManager and ModelManager.