Can mock WifiManager.getScanResults.
TODO: Address the ServiceConnection leaks happening at the Android
platform level. This isn't a _huge_ deal, but we shouldn't have
memory leaks, ideally.

To test the Wifi functionality, we introduce a button to manually
start a scan and insert those results into the db. Then in the
terminal we modify the SSID names to a single phrase and see how
the Settings app reacts. It _works_, but it's not perfect.

Started working on a broadcast receiver for WifiManager events,
so we can fake these in the future.

Modified MockerService to handle wifi clients, and push out mock
scan results (somewhat indefinitely).

MockScanResult was modified to contain a group_id column. This
lets us group scan results as a list so we can send them along
to the platform as a list and we don't have to build it over time.

Renamed MockScanResultManager->MockWifiManager.